+++ www/modules/email/language/nl.php
@@ -311,4 +311,5 @@ Instellingen -> Accounts -> Dubbelklik account -> Mappen.',
 	'Show BCC field by default','Toon BCC veld standaard',
 	'Show CC field by default','Toon CC veld standaard',
 	'Default template','Standaard sjabloon',
+	'Your token has possibly expired. A new window will be opened in which you can renew your token.' => 'Um token is mogelijk verlopen. Er wordt een nieuw venster geopend waarbinnen u uw token kunt vernieuwen.'
 );